requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
tldextract==3.4.4
validators==0.22.0
pyperclip==1.8.2 